Rename module_config.rs to config.rs
