+++ starship/src/lib.rs
@@ -1,9 +1,8 @@
 // Lib is present to allow for benchmarking
-mod config;
+pub mod config;
 pub mod configs;
 pub mod context;
 pub mod module;
-pub mod module_config;
 pub mod modules;
 pub mod print;
 pub mod segment;
